ml-service/requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
numpy
scikit-learn
yfinance
python-dotenv
firecrawl-py
openai
